Refactor: rename City to Cities::City.
- renames `City` class to `Cities::City` to keep every class in `Cities::` namespace (`City` is too general)
- moves static methods from `City` to `Cities`
