+++ lib/cities/country.rb
@@ -3,11 +3,11 @@ module ISO3166
   class Country
 
     def cities
-      City.cities_in_country(alpha2)
+      Cities.cities_in_country(alpha2)
     end
 
     def cities?
-      City.cities_in_country?(alpha2)
+      Cities.cities_in_country?(alpha2)
     end
 
   end
